Minor Changes and adjustments to constants
Though I changed constants according to Hannah's code, I think the values of speeds need further tuning. The VL and VR should not be too low while TVL and TVR should not be too high.
Also, try to use the constant as they were or at most devide them by two to achieve a difference that is equal to a constant. Try to avoid deviding by 3 or other numbers as that defeats the purpose of having constants. If changes were to made, please adjust the constants directly or add a limiting function to control its min / max values.